refactor: move main loop into game engine class

+++ assignment-01/src/main/java/pcd/sketch01/V2d.java
@@ -8,11 +8,11 @@ public record V2d(double x, double y)  {
     }
 
     public double abs(){
-        return (double)Math.sqrt(x*x+y*y);
+        return Math.sqrt(x*x+y*y);
     }
 
     public V2d getNormalized(){
-        double module=(double)Math.sqrt(x*x+y*y);
+        double module= Math.sqrt(x*x+y*y);
         return new V2d(x/module,y/module);
     }
